+++ addon/components/flexberry-objectlistview.js
@@ -24,6 +24,102 @@ export default FlexberryBaseComponent.extend({
   _showFilters: Ember.computed.oneWay('filters'),
 
   /**
+    Store the action name at controller for loading records.
+
+    @property _loadRecords
+    @type String
+    @default 'loadRecords'
+    @private
+  */
+  _loadRecords: 'loadRecords',
+
+  /**
+    Store the action name at controller for switch to the hierarchical mode.
+
+    @property _switchHierarchicalMode
+    @type String
+    @default 'switchHierarchicalMode'
+    @private
+  */
+  _switchHierarchicalMode: 'switchHierarchicalMode',
+
+  /**
+    Store the action name at controller for save the hierarchical attribute name.
+
+    @property _saveHierarchicalAttribute
+    @type String
+    @default 'saveHierarchicalAttribute'
+    @private
+  */
+  _saveHierarchicalAttribute: 'saveHierarchicalAttribute',
+
+  /**
+    Flag indicate when available the hierarchical mode.
+
+    @property _availableHierarchicalMode
+    @type Boolean
+    @default false
+    @private
+  */
+  _availableHierarchicalMode: false,
+
+  /**
+    Flag indicate when component is in the hierarchical mode.
+
+    @property _inHierarchicalMode
+    @type Boolean
+    @default false
+    @private
+  */
+  _inHierarchicalMode: Ember.computed(function() {
+    return this.get('currentController').get('inHierarchicalMode');
+  }),
+
+  /**
+    Store the attribute name set by `hierarchyByAttribute`.
+
+    @property _hierarchicalAttribute
+    @type String
+    @private
+  */
+  _hierarchicalAttribute: undefined,
+
+  /**
+    Set the attribute name to hierarchy build.
+    If specified, will attempt to build on this attribute hierarchy.
+
+    @property hierarchyByAttribute
+    @type String
+  */
+  hierarchyByAttribute: Ember.computed({
+    get() {
+      return this.get('_hierarchicalAttribute');
+    },
+    set(key, value) {
+      this.set('_hierarchicalAttribute', value);
+      this.sendAction('_saveHierarchicalAttribute', value, true);
+      return value;
+    },
+  }),
+
+  /**
+    Indent to indicate hierarchy, can be used HTML.
+
+    @property hierarchicalIndent
+    @type String
+  */
+  hierarchicalIndent: undefined,
+
+  /**
+    Flag used for disable the hierarchical mode.
+
+    @property disableHierarchicalMode
+    @type Boolean
+    @default false
+  */
+  disableHierarchicalMode: false,
+
+  /**
     Text to be displayed in table body, if content is not defined or empty.
 
     @property placeholder
@@ -552,7 +648,39 @@ export default FlexberryBaseComponent.extend({
     filterByAnyMatch(pattern) {
       throw new Error('No handler for filterByAnyMatch action set for flexberry-objectlistview. ' +
                       'Set handler like {{flexberry-objectlistview ... filterByAnyMatch=(action "filterByAnyMatch")}}.');
-    }
+    },
+
+    /**
+      Set availability hierarchical mode, and save the attribute name in controller.
+
+      @method actions.availableHierarchicalMode
+      @param {String} hierarchicalAttribute Attribute name to hierarchy building.
+    */
+    availableHierarchicalMode(hierarchicalAttribute) {
+      this.toggleProperty('_availableHierarchicalMode');
+      this.sendAction('_saveHierarchicalAttribute', hierarchicalAttribute);
+    },
+
+    /**
+      Called controller action to switch in hierarchical mode.
+
+      @method actions.switchHierarchicalMode
+    */
+    switchHierarchicalMode() {
+      this.sendAction('_switchHierarchicalMode');
+    },
+
+    /**
+      Redirects the call to controller..
+
+      @method actions.loadRecords
+      @param {String} Primary key.
+      @param {ObjectListViewRowComponent} Instance of {{#crossLink "ObjectListViewRowComponent"}}{{/crossLink}}.
+      @param {String} Property name.
+    */
+    loadRecords(id, target, property) {
+      this.sendAction('_loadRecords', id, target, property);
+    },
   },
 
   /**