+++ addon/services/log.js
@@ -229,8 +229,10 @@ export default Ember.Service.extend({
     this._super(...arguments);
 
     let _this = this;
+    let originalEmberLoggerError = Ember.Logger.error;
     let onError = function(error, rethrowError) {
       let message = error.message || error.toString();
+
       let formattedMessage = JSON.stringify(Ember.merge({
         name: null,
         message: null,
@@ -238,17 +240,26 @@ export default Ember.Service.extend({
         lineNumber: null,
         columnNumber: null,
         stack: null
-      }, error));
+      }, {
+        name: error.name,
+        message: error.message,
+        fileName: error.fileName,
+        lineNumber: error.lineNumber,
+        columnNumber: error.columnNumber,
+        stack: error.stack
+      }));
 
       _this._storeToApplicationLog(messageCategory.error, message, formattedMessage);
 
       if (rethrowError === false) {
+
         // Break execution if rethrowError === false.
         return;
       } else {
+
         // Rethrow an error, because Ember.onerror handler has no bubbling
         // and stored error won't appear in browser's console without rethrowing.
-        throw error.stack ? error : (error.message ? error.message : error);
+        originalEmberLoggerError(...arguments);
       }
     };
 
@@ -258,7 +269,6 @@ export default Ember.Service.extend({
     Ember.RSVP.on('error', onError);
 
     // Extend Ember.Logger.log logic.
-    let originalEmberLoggerError = Ember.Logger.error;
     Ember.Logger.error = function() {
       originalEmberLoggerError(...arguments);
 
@@ -348,9 +358,12 @@ export default Ember.Service.extend({
       return;
     }
 
-    store.createRecord(applicationLogModelName, applicationLogProperties).save().catch(() => {
-      // Switch off remote logging on rejection to avoid infinite loop.
-      this.set('enabled', false);
-    });
+    return store.createRecord(applicationLogModelName, applicationLogProperties).save().
+      then(
+        result => {return result;},
+
+        // Switch off remote logging on rejection to avoid infinite loop.
+        reason => {this.set('enabled', false);}
+      );
   },
 });